Melee Drawer Fix, Soul Orb Straight Line, Orb Pattern Pyramid
In this build I made an actual orb pattern which consists of an
alternating pyramid using soul orbs which fire in a straight line. I
also fixed the melee drawer after noticing a problem with it. Play in
Editor gives slightly different results than standalone and separate
window which I will investigate soon.